# src/backend/app/models/todo.py
from sqlmodel import SQLModel, Field, Relationship
from datetime import datetime
import uuid
from typing import TYPE_CHECKING, Optional
if TYPE_CHECKING:
    from .user import User


class TodoBase(SQLModel):
    title: str
    description: Optional[str] = None
    completed: bool = Field(default=False, nullable=False)


class Todo(TodoBase, table=True):
    id: uuid.UUID = Field(default_factory=uuid.uuid4, primary_key=True)
    user_id: uuid.UUID = Field(foreign_key="user.id", nullable=False)
    created_at: datetime = Field(default_factory=datetime.utcnow)

    # Relationship
    user: Optional["User"] = Relationship(back_populates="todos")